old_version/duration_prediction/requirements.txt
pydantic>=2.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.20.0
networkx
scikit-learn
chromadb
